Add utils' site, installations submodules
