+++ utils/__init__.py
@@ -1,3 +1,5 @@
 from .net import *  # noqa: F401, F403
 from .types import *  # noqa: F401, F403
 from .utils import *  # noqa: F401, F403
+from .site import *  # noqa: F401, F403
+from .installations import *  # noqa: F401, F403
